W445: +2 MATCH, first-party nd -1,476; h_camp finally moves
Gate run: 14,055 rows. First-party MATCH 6,903 -> 6,905 (83.8%); first-party
normalized_diff 1,204,264 -> 1,202,788 (-1,476) across 33 changed functions.
Zero MATCHes lost, zero regressions, zero new over-window (3,447 unchanged).
92 tests pass. Full build reproduces retail SLUS_216.21 sha1
3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb. The MATCH-class relocation audit is
still clean at 0 functions across 22,165 call relocations.

Eighteen lanes on disjoint file sets, all working the prologue saved-register
class established in W444.

New MATCHes, both in datCalc.c

  00301540  FUN_00301540   7 -> 0   324/336
  00306bc0  FUN_00306bc0   1 -> 0   456/464

`00306bc0` is the notable one: it had been on the confirmed-floor list for many
waves as an unbreakable commutative `addu` operand order. The fix is the one
`skill://mwccps2-operand-order-inline-helper` prescribes and nobody had tried
here — a `static inline u32 datCalcAddIndexBase(u32 index, u32 base)` called with
the index first, so the operands reach MWCC in retail's order. `00301540`
collapsed two redundant locals into a single `u16 offset` assigned identically at
both sites.

Largest nd reductions

  0039b950  mt_evtCustomEvent.c  FUN_0039b950              -462
  001b3480  k_fldrc.c            FUN_001b3480     931 -> 728  (-203)
  003a05b0  mt_evtMessage.c      FUN_003a05b0              -141
  0016af90  h_camp.c             FUN_0016af90    1912 -> 1819  (-93)
  0038cc10  mt_evtMisc.c         FUN_0038cc10               -73
  004281f0  y_misc.c             FUN_004281f0    5012 -> 4965  (-47)
  003a0960  mt_evtMessage.c      FUN_003a0960     311 -> 264   (-47)
  00359e50  primitive.c          primCylinderLine3D 291 -> 250 (-41)
  0043bda0  y_misc.c             FUN_0043bda0     295 -> 255   (-40)
  00227f30  bp_tex.c             FUN_00227F30    1522 -> 1483  (-39)
  003abb10  itfMesManager.c      FUN_003abb10    1239 -> 1201  (-38)
  001a8140  k_clump.c            func_001a8140   1079 -> 1045  (-34)

`FUN_001b3480` is the wave's cleanest demonstration: removing cached `id` and
`mode` locals and reloading `query[1]` / `query[3]` directly after the calls that
produce them took the saved set from ours `{s0-s6,ra}` to retail's exact
`{s0-s4,ra}`.

h_camp moves for the first time in four waves

`FUN_0016af90` went 1,912 -> 1,819 and reached retail's EXACT frame (-0x1d0) and
EXACT saved-register set, via a three-part chain:

  1. a materialised `parent` pointer where the source passed NULL to the draw call
  2. `recordOffset = recordIndex * 0x24` computed once, with every record field
     reached through `param4 + recordOffset` — this is what forces s6 and s7 live
  3. a hoisted `inverseAlpha = 0xffU - (u32)param_5` reused across text and digits

Step 2 is the transferable one. Three more rows carry the identical missing set
`s5, s6, s7, fp` — `0013afd0` (2,095), `00139fc0` (2,793), `0015b430` (2,848) —
and two of those already have retail's frame, so they are clean tests for the
next wave.

Three h_camp hypotheses were tested to destruction and are now closed:

  - **Coordinates as integers.** Retail `00148c10` builds 379 with
    `addiu`/`mtc1`/`cvt.s.w`, so it looked like the source held integers. Every
    form measured neutral: `(f32)379`, an int local, a hex cast, sharing one
    integer across both pair words (worse at 1,941). Note retail's y comes from
    `lui 0x4140` — a genuine float constant — so the two halves differ in origin.
  - **Address-taken scratch.** Forcing the pair through `CampPair *p = &pair`
    measured worse (1,910 vs 1,906).
  - **Contiguous scratch array.** Replacing the 22 separate `CampBits` locals
    with `CampBits scratch[22]` was exactly neutral.

Also confirmed: `func_0018bc10`'s depth parameter genuinely is a float in `$f12`
and `100.0f` is materialised as `lui 0x42c8` — its prototype must not change.

The integer-materialisation mechanism, and its boundary

VpadW445 landed k_sceneDraw `func_001a2a80` 3,008 -> 2,988 by changing a helper's
`floatDelta` parameter from `f32` to `s32`, writing the additions as
`(f32)floatDelta`, and converting the call-site literals to integers. That
removed a surplus f20 and moved the frame -0xe0 -> -0xd0. Both halves were
needed; the literals alone are not enough.

The discriminator, cheap to apply, is in the retail decode of the call site:
`lui <float bits>` then `mtc1` means a genuine float constant, leave it alone;
`addiu $reg, $zero, <small int>` then `mtc1` then `cvt.s.w` means the source held
an integer. mdlEffect audited its FPR rows this way and found NO applicable site
— every constant there is `lui` + `mtc1` — so it is a real mechanism with a
narrow domain, not a universal one.

Rejected in review

  - y_timeLimit `0045c8c0` had a real statement deleted during probing
    (`fVar11 = 1.0f + fVar12;`). It survived its lane's scoped verify because the
    lane measured a different function. Restored; the function is back at 570
    from 840, and the wave's net moved from -1,206 to -1,476.
  - The three optimisation-pragma probes flagged in W444 stayed reverted; the one
    pragma occurrence touched this wave (mdlEffect `0032c340`) predates the wave.

Tooling correction

`tools/reloc_audit.py` now defaults to `--status MATCH` and documents why.
W443's headline "990 functions with a wrong call target" was WRONG and is
withdrawn: `retail_target` is read at our object's relocation OFFSET, which only
names the same instruction in both binaries when the function already matches. In
a NONMATCHING function the offsets have shifted and the comparison returns
nonsense — impossible targets such as `0x01080060`. The MATCH-class finding
(23 functions, all fixed) stands and is unaffected.

Zero-gain lanes, closed at exact baseline with measured negatives: Calendar,
BrPanel and the mdlEffect FPR audit.

+++ src/Event/Fcl/fclCombineList.c
@@ -1700,8 +1700,8 @@ void fclCombineList003dd800(s32 x, s32 y, s32 alpha, FclOwner* owner,
     record.candidate_data = candidate_data;
     record.owner = owner;
     record.candidate = candidate;
-    FUN_003e0680((u64)(u32)(void *)candidate_data->callback_target,
-                  (code*)fclCombineList003dd260, (u64)(u32)(void *)&record);
+    FUN_003e0680((u32)(void *)candidate_data->callback_target,
+                  (code*)fclCombineList003dd260, (u32)(void *)&record);
 }
 
 // FUN_003dd9e0
